fix: Moving of some spec files to src root folder
Several spec files was in __tests__ folder

Now specs files are at src/ and the linkeds
snapshots in __snapshots__ folder
