+++ packages/cozy-stack-client/src/CozyStackClient.spec.js
@@ -1,13 +1,13 @@
 /* eslint-env jest */
 /* global fetch */
 
-import CozyStackClient from '../CozyStackClient'
-import DocumentCollection from '../DocumentCollection'
-import JobCollection from '../JobCollection'
-import KonnectorCollection from '../KonnectorCollection'
+import CozyStackClient from './CozyStackClient'
+import DocumentCollection from './DocumentCollection'
+import JobCollection from './JobCollection'
+import KonnectorCollection from './KonnectorCollection'
 import jestFetchMock from 'jest-fetch-mock'
-import AppToken from '../AppToken'
-import { FetchError } from '../errors'
+import AppToken from './AppToken'
+import { FetchError } from './errors'
 
 const FAKE_RESPONSE = {
   offset: 0,
